refactor(business): 12.3 主动发起识别成功后，更新语音识别状态为：生成字幕中（SP）
- 将 transStatusCode 字段类型从 Long 改为 Integer
- 实现语音识别请求的详细处理逻辑
- 添加异常处理和日志记录
- 更新数据库记录以反映语音识别状态

+++ ruoyi-common/ruoyi-common-dependency/src/main/java/org/dromara/common/dependency/business/domain/BizFiletrans.java
@@ -89,7 +89,7 @@ public class BizFiletrans extends TenantEntity {
     /**
      * 转换状态码
      */
-    private Long transStatusCode;
+    private Integer transStatusCode;
 
     /**
      * 转换状态说明
